requirements.txt
alabaster==0.7.6
amqp==1.4.8
anyjson==0.3.3
-e git+https://readonlygetgoing:gg11jenkins@github.com/getgoing/atlas.git@73305bb4f71ba1e365db74610197be024e3e721d#egg=atlas
Babel==2.1.1
behave==1.2.6.dev0
billiard==3.3.0.22
-e git+https://readonlygetgoing:gg11jenkins@github.com/getgoing/boeing.git@3a2165c88a546173320143f880990651d27baf13#egg=boeing
boto==2.38.0
celery==3.1.19
certifi==2015.4.28
ciso8601==1.0.1
click==6.2
contextlib2==0.4.0
-e git+https://readonlygetgoing:gg11jenkins@github.com/getgoing/courtyard.git@f34f854c450a109077e119831c019f216bccdf07#egg=courtyard
coverage==4.0.3
crcmod==1.7
decorator==4.0.6
deepdiff==0.6.0
diff-match-patch==20121119
dj-database-url==0.3.0
Django==1.8.7
django-annoying==0.8.4
django-celery==3.1.17
django-countries==3.4.1
django-dotenv==1.3.0
django-extensions==1.5.9
django-grappelli==2.7.2
django-localflavor==1.2
django-mptt==0.7.4
django-nose==1.4.2
django-storages==1.1.8
django-wkhtmltopdf==3.0.0
djangorestframework==3.0.5
djorm-ext-core==0.7
docopt==0.6.2
docutils==0.12
drf-nested-routers==0.11.1
ecdsa==0.13
elasticsearch==2.1.0
elasticsearch-dsl==0.0.9
enum34==1.1.1
eventlet==0.17.4
factory-boy==2.6.0
fake-factory==0.5.3
-e git+https://readonlygetgoing:gg11jenkins@github.com/getgoing/fasttrace.git@fcd1613c2d63802366d8d9352ccb6e9aedfecf9c#egg=fasttrace
faulthandler==2.4
first==2.0.1
flake8==2.5.1
flake8-docstrings==0.2.1.post1
flake8-import-order==0.5.3
flake8-quotes==0.0.1
funcsigs==0.4
functools32==3.2.3.post2
funcy==1.6
futures==3.0.3
gnureadline==6.3.3
greenlet==0.4.9
gunicorn==19.4.1
hiredis==0.2.0
honcho==0.6.6
httplib2==0.9.1
-e git+https://readonlygetgoing:gg11jenkins@github.com/getgoing/hyatt.git@4b086f8fcb2df14ec3d0676988c1b7945de9f777#egg=hyatt
hypothesis==1.16.0
icalendar==3.9.0
inflection==0.3.1
ipython==3.0.0
Jinja2==2.8
jsl==0.1.2
jsonpath-rw==1.4.0
jsonschema==2.5.1
-e git+https://readonlygetgoing:gg11jenkins@github.com/getgoing/junkers.git@c7dd2b3aeb34b8d1f498ea4347d96435e99781d6#egg=junkers
kombu==3.0.30
kyotocabinet==1.9
librabbitmq==1.6.1
lru-dict==1.1.3
lxml==3.5.0
lz4==0.7.0
MarkupSafe==0.23
mccabe==0.3.1
-e git+https://readonlygetgoing:gg11jenkins@github.com/getgoing/mcdonnell.git@c755866c3655111c8b7a7e59943e412b8fd90eff#egg=mcdonnell
mock==1.3.0
msgpack-python==0.4.6
newrelic==2.58.2.45
nose==1.3.7
paramiko==1.16.0
parse==1.6.6
parse-type==0.3.4
pbr==1.8.1
pep257==0.5.0
pep8==1.5.7
pika==0.10.0
Pillow==3.0.0
pip-tools==1.3.0
ply==3.8
polib==1.0.3
progressbar==2.3
psycopg2==2.6.1
pycrypto==2.6.1
pyes==0.99.6
pyflakes==1.0.0
Pygments==2.0.2
pylibmc==1.5.0
pyparsing==2.0.6
python-dateutil==2.4.2
python-magic==0.4.10
pytz==2015.7
PyYAML==3.11
pyzmq==14.5.0
-e git+https://readonlygetgoing:gg11jenkins@github.com/getgoing/quicksilver.git@3dbe6ee50997383ab450bda1fb94ac25ea514ba0#egg=quicksilver
-e git+https://readonlygetgoing:gg11jenkins@github.com/getgoing/radisson.git@ac78e3d9ab54c2a1d04ebecdae520944a68ad3c9#egg=radisson
raven==5.9.0
redis==2.10.5
regex==2015.11.22
repoze.lru==0.6
requests==2.8.1
requests-toolbelt==0.5.0
Routes==2.2
rules==1.1.1
s3cmd==1.6.0
sendgrid==0.1.4
simplejson==3.8.1
six==1.10.0
snowballstemmer==1.2.0
Sphinx==1.3.1
sphinx-rtd-theme==0.1.9
SQLAlchemy==1.0.9
-e git+https://readonlygetgoing:gg11jenkins@github.com/getgoing/starwood.git@52a4d263825c4776c5ffb0e1d99f734c313fe2a6#egg=starwood
stevedore==1.3.0
-e git+https://github.com/ventura/suds-gzip.git@36f6afd15520436ec942c2e9d1edcaca08b1551f#egg=suds
-e git+https://readonlygetgoing:gg11jenkins@github.com/getgoing/tune.git@2ce118a4da9ce1fd2d4a91c0a5ce95632468709a#egg=tune
ujson==1.34
Unidecode==0.4.18
urllib3==1.12
virtualenv==12.0.7
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.2
xlrd==0.9.4
XlsxWriter==0.7.7
xlwt==1.0.0
xmltodict==0.9.2
